requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.51
botocore==1.27.51
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-pivot==1.9.0
django-storages==1.13.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11